+++ src/factories/__init__.py
@@ -1,0 +1,11 @@
+"""
+Factory classes for creating components from configuration.
+"""
+
+from .model import ModelFactory, create_mistral, create_dialogpt
+
+__all__ = [
+    "ModelFactory",
+    "create_mistral", 
+    "create_dialogpt"
+]