+++ src/utils/__init__.py
@@ -1,0 +1,11 @@
+# src/utils/__init__.py
+"""
+Utility functions and classes
+"""
+
+from .logger import Logger, create_logger
+
+__all__ = [
+    "Logger",
+    "create_logger"
+]